Pull st0 EntityRelicOrb to .h for git diffing
